initial refactoring of pretty modules
