+++ src/Cooked/Attack/DatumHijacking.hs
@@ -12,7 +12,7 @@ where
 
 import Control.Monad
 import Cooked.Output
-import Cooked.Pretty.Class
+import Cooked.Pretty.Common
 import Cooked.Skeleton
 import Cooked.Tweak
 import Data.Maybe
