+++ src/Cooked/Pretty/Common.hs
@@ -2,26 +2,54 @@
 -- cooked-validators
 module Cooked.Pretty.Common
   ( DocCooked,
+    PrettyCooked (..),
+    printCookedOpt,
+    printCooked,
     renderString,
     prettyItemize,
     prettyItemizeNoTitle,
     prettyItemizeNonEmpty,
     prettyEnumerate,
+    prettyCooked,
     prettyHash,
   )
 where
 
-import Cooked.Pretty.Options (PrettyCookedHashOpts (..))
+import Cooked.Pretty.Hashable
+import Cooked.Pretty.Options
 import Data.ByteString qualified as ByteString
+import Data.Default
 import Data.Map qualified as Map
+import Data.Ratio
 import Numeric qualified
 import PlutusTx.Builtins.Internal qualified as PlutusTx
 import Prettyprinter (Doc, (<+>))
 import Prettyprinter qualified as PP
 import Prettyprinter.Render.String qualified as PP
+import Prettyprinter.Render.Text qualified as PP
 
 -- | A standard 'PP.Doc' without any annotation
 type DocCooked = Doc ()
+
+-- | Type class of things that can be pretty printed using options
+class PrettyCooked a where
+  -- | Pretty prints an element based on some 'PrettyCookedOpts'
+  prettyCookedOpt :: PrettyCookedOpts -> a -> DocCooked
+
+-- | Pretty prints an element without requiring some options
+prettyCooked :: (PrettyCooked a) => a -> DocCooked
+prettyCooked = prettyCookedOpt def
+
+-- | Use this in the REPL as an alternative to the default 'print' function when
+-- dealing with pretty-printable cooked values.
+--
+-- For example, @printCookedOpt def runMockChain i0 foo@
+printCookedOpt :: (PrettyCooked a) => PrettyCookedOpts -> a -> IO ()
+printCookedOpt opts e = PP.putDoc $ prettyCookedOpt opts e <+> PP.line
+
+-- | Version of 'printCookedOpt' that uses default pretty printing options.
+printCooked :: (PrettyCooked a) => a -> IO ()
+printCooked = printCookedOpt def
 
 -- | Use this to convert a pretty-printer to a regular show function using
 -- default layout options. This is used in "Testing" because Tasty uses strings.
@@ -60,27 +88,61 @@ prettyEnumerate title bullet items =
         zipWith (\index item -> PP.pretty index <> bullet <+> PP.align item) [1 :: Int ..] items
     ]
 
--- | Pretty print a prefix of a hash with a given length.
-prettyHash :: PrettyCookedHashOpts -> PlutusTx.BuiltinByteString -> DocCooked
-prettyHash PrettyCookedHashOpts {..} bbs@(PlutusTx.BuiltinByteString bs) =
-  let hexRepresentation :: DocCooked
-      hexRepresentation =
-        "#"
-          <> ( PP.pretty
-                 . take pcOptHashLength
-                 . concatMap
-                   ( \x ->
-                       let res = Numeric.showHex x ""
-                        in if length res == 1
-                             then '0' : res
-                             else res
-                   )
-                 . ByteString.unpack
-             )
-            bs
-   in case Map.lookup bbs pcOptHashNames of
-        Nothing -> hexRepresentation
-        Just name ->
-          if pcOptHashVerbose
-            then hexRepresentation <+> PP.parens (PP.pretty name)
-            else PP.pretty name
+instance {-# OVERLAPPABLE #-} (PrettyCooked a) => PrettyCooked [a] where
+  prettyCookedOpt opts = prettyItemizeNoTitle "-" . map (prettyCookedOpt opts)
+
+-- * Pretty printing of hashable data types
+
+-- | Pretty prints hashable elements based on 'pcOptHashes' in the
+-- 'PrettyCookedOpts'. This cannot be made an instance as it would be
+-- undecidable (the hope was @(ToHash a) => PrettyCooked a@)
+prettyHash :: (ToHash a) => PrettyCookedOpts -> a -> DocCooked
+prettyHash
+  (PrettyCookedOpts {pcOptHashes = PrettyCookedHashOpts {..}})
+  (toHash -> bbs@(PlutusTx.BuiltinByteString bs)) =
+    let hexRepresentation =
+          PP.pretty
+            . take pcOptHashLength
+            . concatMap
+              -- We pad the result of 'Numeric.showHex' to reach exactly
+              -- 2 characters as it might only have 1 in some occasions.
+              (\((`Numeric.showHex` "") -> res) -> if length res == 1 then '0' : res else res)
+            . ByteString.unpack
+            $ bs
+     in case Map.lookup bbs pcOptHashNames of
+          Nothing -> "#" <> hexRepresentation
+          Just name | pcOptHashVerbose -> "#" <> hexRepresentation <+> PP.parens (PP.pretty name)
+          Just name -> PP.pretty name
+
+-- * Pretty instances for some common base types
+
+instance PrettyCooked Int where
+  prettyCookedOpt _ = PP.pretty
+
+instance PrettyCooked Integer where
+  prettyCookedOpt opts =
+    if pcOptNumericUnderscores opts
+      then prettyNumericUnderscore
+      else PP.pretty
+    where
+      -- prettyNumericUnderscore 23798423723
+      -- 23_798_423_723
+      prettyNumericUnderscore :: Integer -> DocCooked
+      prettyNumericUnderscore i
+        | 0 == i = "0"
+        | i > 0 = psnTerm "" 0 i
+        | otherwise = "-" <> psnTerm "" 0 (-i)
+        where
+          psnTerm :: DocCooked -> Integer -> Integer -> DocCooked
+          psnTerm acc _ 0 = acc
+          psnTerm acc 3 nb = psnTerm (PP.pretty (nb `mod` 10) <> "_" <> acc) 1 (nb `div` 10)
+          psnTerm acc n nb = psnTerm (PP.pretty (nb `mod` 10) <> acc) (n + 1) (nb `div` 10)
+
+instance PrettyCooked Bool where
+  prettyCookedOpt _ = PP.pretty
+
+instance PrettyCooked () where
+  prettyCookedOpt _ = PP.pretty
+
+instance PrettyCooked Rational where
+  prettyCookedOpt opts q = "(" <+> prettyCookedOpt opts (numerator q) <+> "/" <+> prettyCookedOpt opts (denominator q) <+> ")"
