+++ tests/Cooked/Attack/DatumHijackingSpec.hs
@@ -27,7 +27,7 @@ import Test.Tasty.HUnit
 data LockDatum = FirstLock | SecondLock deriving (Show, Eq)
 
 instance PrettyCooked LockDatum where
-  prettyCooked = viaShow
+  prettyCookedOpt _ = viaShow
 
 instance PlutusTx.Eq LockDatum where
   {-# INLINEABLE (==) #-}
